Security update
Made sure it is not possible to forcefully get info about strikes that do no belong to the user.
Also adjusted the Strike find method from Integer to Long for the id's are of type Long.
Also added the check for url on finding articles.

+++ app/models/Strike.java
@@ -380,7 +380,7 @@ public class Strike extends Model{
     /**
      * Method to find the strikes in the database
      */
-    public static Finder<Integer, Strike> find = new Finder<>(Strike.class);
+    public static Finder<Long, Strike> find = new Finder<>(Strike.class);
 
     /**
      * gets all strikes from the database
